Set kubeReserved defaults according to instance type
This change implements default values for `kube-reserved` the same way
as the [official AMIs] do which is dependent on the instance type.

Reserved memory and cpu are set according to a formula while the storage
is always set to 1GiB. See the source code for details.

Instance type information (such as available storage, memory, and cpu)
may not only change in the future but also new instances will be added.
There is a make target `generate-kube-reserved` which fetches all instance
types by querying the AWS API `ec2:DescribeInstanceTypes` in each
region. It then saves it to `pkg/nodebootstrap/reserved_data.go` ready
to be taken as input for computing appropriate default values.

[official AMIs]: https://github.com/awslabs/amazon-eks-ami/blob/master/files/bootstrap.sh